copy 2 files and update 3 files
